Reorganize benchmarks
