+++ Cargo.toml
@@ -54,21 +54,6 @@ harness = false
 name = "li"
 harness = false
 
-[[bench]]
-name = "ho_box"
-harness = false
-required-features = ["alloc"]
-
-[[bench]]
-name = "hi_box"
-harness = false
-required-features = ["alloc"]
-
-[[bench]]
-name = "li_box"
-harness = false
-required-features = ["alloc"]
-
 [[example]]
 name = "gpt_async"
 required-features = ["std"]
